expand items available +add rarity + debugger console

+++ Assets/Scripts/Combat/IDamageable.cs
@@ -3,6 +3,6 @@ namespace CryptKnight.Combat
     public interface IDamageable
     {
         DamageableTarget TargetType { get; }
-        void ApplyDamage(int damage);
+        void ApplyDamage(float damage);
     }
 }
